fix: No longer display duplicate issues when a user uploads many files at once with the same issues.
feat: added a duplicate file check, just so a user doesn't accidently try uploading the same file twice. We purposefully allow a user to upload files with the same name as long as they are not the exact same size (in bytes) to reduce user friction. I will have a new feature in the future to warn users about duplicate Invoice ID's if they exist and allow users to resolve duplicate invoice ID's.