feat: Update key for uploadedfiles hash map after we receive a response from the server with new unique ID "fileName". This replaces the temporary clientID that's used during optimistic loading.
fix: erase e.target.value in handleClickToBrowse so that the same file can be uploaded if the user decides to remove it from the list of files to upload then try to re-add. Before this fix if the user attempted to do so with the click to browse feature it would not allow you to do so until you uploaded a different file.